Updates how conditional fields are shown/hidden.
